pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "richyplayer"
version = "1.0.0"
dependencies = [
  "pygame-ce>=2.5.3",
  "numpy>=2.2.3",
  "opencv-python>=4.11.0.86",
  "pygbag>=0.9.2",
]
requires-python = ">=3.12"
authors = [
  {name = "Richmond Kleinson"},
]
maintainers = [
  {name = "Richmond Kleinson"}
]
description = "Video player for desktop & web"
readme = "README.md"
license = { file = "LICENSE" }
keywords = ["richyplayer", "video player", "pygbag"]
classifiers = [
  # How mature is this project? Common values are
  #   3 - Alpha
  #   4 - Beta
  #   5 - Production/Stable
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.12",
]

[project.urls]
Homepage = "https://github.com/richkdev/richyplayer"
Documentation = "https://github.com/richkdev/richyplayer/blob/main/docs"
Repository = "https://github.com/richkdev/richyplayer.git"
"Bug Tracker" = "https://github.com/richkdev/richyplayer/issues"
Changelog = "https://github.com/richkdev/richyplayer/blob/master/CHANGELOG.md"
